+++ data-raw/data-raw.R
@@ -1,8 +1,3 @@
-## Get up-to-date stations data
-withr::with_options(list("weathercan.verbosity" = "verbose"), {
-  stations_dl_internal()
-})
-
 kamloops_day <- weather_dl(
   51423,
   start = "2016-01-01",